+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [2.0.1](https://github.com/witchesofthehill/manabrew/compare/v2.0.0...v2.0.1) (2026-07-24)
+
+### Fixes
+
+* **ironsmith:** use protocol v1 bridge ([#533](https://github.com/witchesofthehill/manabrew/issues/533)) ([e5053d9](https://github.com/witchesofthehill/manabrew/commit/e5053d9e0a671dccca23436b652c541aef0f92c4))
+
+## [2.0.0](https://github.com/witchesofthehill/manabrew/compare/v1.14.0...v2.0.0) (2026-07-24)
+
+### Features
+
+* **protocol:** add source card metadata to prompts ([#522](https://github.com/witchesofthehill/manabrew/issues/522)) ([1dfadc5](https://github.com/witchesofthehill/manabrew/commit/1dfadc59bf687aa2162deb84667ca080f6814e83))
+
 ## [1.14.0](https://github.com/witchesofthehill/manabrew/compare/v1.13.1...v1.14.0) (2026-07-23)
 
 ### Features